Break out training config
