+++ src/bin/train_tokenizer.rs
@@ -1,3 +1,5 @@
+pub mod train;
+
 use clap::Parser;
 use std::{collections::HashMap, env, path::PathBuf, vec};
